refactor: reorganize swap components and state management for enhanced clarity and functionality
