+++ clients/desktop/package.json
@@ -42,7 +42,6 @@
     "@lib/utils": "workspace:*",
     "@tanstack/query-core": "^5.69.0",
     "@wailsapp/runtime": "workspace:^",
-    "react-circular-progressbar": "^2.2.0",
     "styled-components": "^6.1.16"
   }
 }
