+++ clients/desktop/src/vault/components/VaultPrimaryActions/index.tsx
@@ -2,12 +2,12 @@ import { CoinKey } from '@core/chain/coin/Coin'
 import { swapEnabledChains } from '@core/chain/swap/swapEnabledChains'
 import { SendPrompt } from '@core/ui/vault/send/SendPrompt'
 import { useCurrentVaultNativeCoins } from '@core/ui/vault/state/currentVaultCoins'
+import { SwapPrompt } from '@core/ui/vault/swap/components/SwapPrompt'
 import { UniformColumnGrid } from '@lib/ui/css/uniformColumnGrid'
 import { ValueProp } from '@lib/ui/props'
 import { isEmpty } from '@lib/utils/array/isEmpty'
 import { isOneOf } from '@lib/utils/array/isOneOf'
 
-import { SwapPrompt } from '../../swap/components/SwapPrompt'
 import { DepositPrompt } from '../DepositPrompts'
 
 export const VaultPrimaryActions = ({ value }: Partial<ValueProp<CoinKey>>) => {
