+++ clients/desktop/wailsjs/runtime/package.json
@@ -4,8 +4,6 @@
   "description": "Wails Javascript runtime library",
   "main": "runtime.js",
   "types": "runtime.d.ts",
-  "scripts": {
-  },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/wailsapp/wails.git"
