+++ core/ui/navigation/routes.tsx
@@ -7,6 +7,7 @@ import { ActiveVaultGuard } from '@core/ui/vault/ActiveVaultGuard'
 import { ManageVaultChainsPage } from '@core/ui/vault/chain/manage'
 import { VaultDetailsPage } from '@core/ui/vault/settings/details'
 import { VaultRenamePage } from '@core/ui/vault/settings/rename'
+import { SwapPage } from '@core/ui/vault/swap/components/SwapPage'
 import { ReactNode } from 'react'
 
 import { SendPage } from '../vault/send/SendPage'
@@ -23,6 +24,7 @@ type SharedPaths = Extract<
   | 'manageVaultChains'
   | 'reshareVault'
   | 'send'
+  | 'swap'
 >
 
 export const sharedRoutes: Record<SharedPaths, ReactNode> = {
@@ -52,6 +54,12 @@ export const sharedRoutes: Record<SharedPaths, ReactNode> = {
   send: (
     <ActiveVaultGuard>
       <SendPage />
+    </ActiveVaultGuard>
+  ),
+
+  swap: (
+    <ActiveVaultGuard>
+      <SwapPage />
     </ActiveVaultGuard>
   ),
 }
