+++ core/ui/vault/swap/components/ChainOption.tsx
@@ -1,5 +1,3 @@
-import { useFromCoin } from '@clients/desktop/src/vault/swap/state/fromCoin'
-import { useToCoin } from '@clients/desktop/src/vault/swap/state/toCoin'
 import { Coin } from '@core/chain/coin/Coin'
 import { isFeeCoin } from '@core/chain/coin/utils/isFeeCoin'
 import { ChainCoinIcon } from '@core/ui/chain/coin/icon/ChainCoinIcon'
@@ -7,6 +5,8 @@ import { getChainEntityIconSrc } from '@core/ui/chain/coin/icon/utils/getChainEn
 import { getCoinLogoSrc } from '@core/ui/chain/coin/icon/utils/getCoinLogoSrc'
 import { shouldDisplayChainLogo } from '@core/ui/chain/coin/icon/utils/shouldDisplayChainLogo'
 import { useTransferDirection } from '@core/ui/state/transferDirection'
+import { useFromCoin } from '@core/ui/vault/swap/state/fromCoin'
+import { useToCoin } from '@core/ui/vault/swap/state/toCoin'
 import { CheckIcon } from '@lib/ui/icons/CheckIcon'
 import { HStack, VStack } from '@lib/ui/layout/Stack'
 import { panel } from '@lib/ui/panel/Panel'
