+++ core/ui/vault/swap/components/SwapCoinInputField/SwapCoinInputField.styled.tsx
