+++ core/ui/vault/swap/form/SwapCoinBalance.tsx
