+++ core/ui/vault/swap/form/amount/ManageFromAmount.tsx
