+++ core/ui/vault/swap/form/balance/SwapCoinBalanceDependant.tsx
