+++ core/ui/vault/swap/form/info/SwapTotalFeeFiatValue.tsx
