+++ core/ui/vault/swap/verify/state/swapTerms.ts
